Added remaining PPU registers + NMI.
This commit finishes adding support for PPU registers, specifically
PPUDATA:
- The buffered reading of lower addresses (it takes 2 reads to get a
  value out in some cases).
- The nametable mirroring behavior, which required that the
  cartridge/mapper classes expose the mirror type.
- The palette area with its special mirroring behavior.

In addition, the framework for each PPU tick has been laid out. The PPU
sets the VBLANK status and sets an NMI for the CPU (which was also added
in this commit) at the appropriate time; it also clears status (VBLANK,
sprite 0 hit, sprite overflow) at the appropriate time.

As always, tests were added to test most of the newly added code.